requirements.txt
Flask
google.generativeai
requests
python.dotenv